// src/types.ts
export enum UserRoles {
  account_owner = "account_owner",
  project_owner = "project_owner",
  contributor = "contributor",
}

export type TokenData = {
  email: string;
  role: UserRoles;
};
// ...
export interface UserAccountView {
  id: number;
  firstName: string;
  lastName: string;
  position: string;
  email: string;
  createdAt: string;
  isBlocked: boolean;
  isDeleted: boolean;
  isLoggedIn: boolean;
  accountId: number;
  accountName: string;
  accountIsActive: boolean;
  accountCreatedAt: string;
  shortCode: string;
  userRole: UserRoles;
  isUserInfoCompleted: boolean;
  isPasswordResetCompleted: boolean;
  isBillingInfoCompleted: boolean;
  isMFAStepCompleted: boolean;
  profilePicUrl?: string;
  isPrimaryAccountOwner: boolean;
  isFirstLogin: boolean;
}